Refactor Progress
- move Progress related classes in separate files
- refactor ConsoleProgressTests to test what the name says
- add comments to ProgressState classes that will eventually be
  deprecated
